admin: move the client app start to typescript
With this, the application does not need `evolution-legacy` anymore for
its execution. It still needs to depend on the package, as some helpers
used by the demo_survey are still in legacy, but that is now just an
internal implementation details. The whole of the app is now in
typescript and future questionnaires should not use legacy anymore.
